+++ tardis/io/model/classic/parse_density.py
@@ -5,6 +5,9 @@ from dataclasses import dataclass
 import numpy as np
 from astropy import units as u
 
+from tardis.io.model.parse_density_configuration import (
+    calculate_density_after_time,
+)
 from tardis.model.mesh import HomologousRadial1DMesh, MeshLocation, MeshQuantity
 
 
@@ -181,20 +184,24 @@ class ExponentialDensitySolver:
 
 @dataclass
 class W7DensitySolver:
-    """W7 Type Ia supernova density profile solver.
+    """W7 Type Ia supernova density profile solver (Branch85 W7).
 
-    This implements the W7 model density profile as a parameterized function.
+    This implements the Branch85 W7 model using a power law density profile
+    with exponent -7, scaled by the time of explosion.
 
     Parameters
     ----------
     time_0 : u.Quantity
-        Reference time for the W7 model.
+        Reference time for the W7 model (default: 0.000231481 day).
     density_0 : u.Quantity
-        Reference density normalization.
+        Reference density normalization (default: 3e29 g/cm^3).
+    velocity_0 : u.Quantity
+        Reference velocity (default: 1 km/s).
     """
 
     time_0: u.Quantity
     density_0: u.Quantity
+    velocity_0: u.Quantity
 
     @classmethod
     def from_config(cls, config):
@@ -213,7 +220,8 @@ class W7DensitySolver:
         density_config = config.model.structure.density
         time_0 = density_config.time_0 * u.Unit(density_config.time_unit)
         density_0 = density_config.rho_0 * u.Unit(density_config.unit)
-        return cls(time_0=time_0, density_0=density_0)
+        velocity_0 = density_config.v_0 * u.Unit(density_config.v_unit)
+        return cls(time_0=time_0, density_0=density_0, velocity_0=velocity_0)
 
     def solve(self, mesh: HomologousRadial1DMesh) -> MeshQuantity:
         """Calculate density for the given mesh.
@@ -229,11 +237,14 @@ class W7DensitySolver:
             Density at cell volumes.
         """
         velocity_middle = 0.5 * (mesh.velocity.data[:-1] + mesh.velocity.data[1:])
-        time_factor = (self.time_0 / mesh.time_explosion) ** 3
-        v_norm = (velocity_middle / (10000 * u.km / u.s)).decompose()
-        velocity_profile = np.exp(-(v_norm**2))
-
-        density_data = self.density_0 * time_factor * velocity_profile
+        
+        # Calculate power law density: rho = density_0 * (v / v_0)^(-7)
+        density_at_time_0 = self.density_0 * (velocity_middle / self.velocity_0) ** (-7)
+        
+        # Scale density from time_0 to time_explosion using calculate_density_after_time
+        density_data = calculate_density_after_time(
+            density_at_time_0, self.time_0, mesh.time_explosion
+        )
 
         return MeshQuantity(
             data=density_data, defined_at=MeshLocation.VOLUME, name="density"
@@ -279,11 +290,9 @@ def parse_density_solver_from_density_config(density_config):
         return ExponentialDensitySolver(density_0=density_0, velocity_0=velocity_0)
     if density_type == "branch85_w7":
         time_0 = density_config.w7_time_0
-        # W7 uses power law with exponent -7 for initial density
         velocity_0 = density_config.w7_v_0
         rho_0 = density_config.w7_rho_0
-        exponent = -7
-        return PowerLawDensitySolver(
-            density_0=rho_0, velocity_0=velocity_0, exponent=exponent
+        return W7DensitySolver(
+            time_0=time_0, density_0=rho_0, velocity_0=velocity_0
         )
     raise ValueError(f"Unrecognized density type '{density_type}'")
